backend/requirements.txt
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djoser==2.1.0
Pillow==8.3.1
PyJWT==2.1.0
requests==2.26.0
psycopg2-binary==2.8.6
python-dotenv==0.19.2
gunicorn==20.1.0
drf-extra-fields==3.2.1
drf-yasg==1.20.0